Fixing the build of TFFI in the new scheme, and fixing the tests

+++ ffiTestLibrary/src/callout.c
@@ -19,7 +19,7 @@ EXPORT(int) sumTwoNumbers(int a, int b){
 }
 
 //Deprecated
-EXPORT(int) sumAFloatAndADouble(float a, double b){
+EXPORT(float) sumAFloatAndADouble(float a, double b){
   return a+b; 
 }
